Remove example for iOS.
