+++ example/ViewPagerExample.js
@@ -4,92 +4,25 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @format
+ * @flow
  */
 
 'use strict';
 
-const React = require('react');
-const ReactNative = require('react-native');
-const {
-  Image,
+import React from 'react';
+import { Image,
   StyleSheet,
   Text,
   TouchableWithoutFeedback,
   TouchableOpacity,
-  View,
-} = ReactNative;
+  View, 
+  Platform } from 'react-native' 
+
 import ViewPagerAndroid from '@react-native-community/viewpager';
-
-const PAGES = 5;
-const BGCOLOR = ['#fdc08e', '#fff6b9', '#99d1b7', '#dde5fe', '#f79273'];
-const IMAGE_URIS = [
-  'https://apod.nasa.gov/apod/image/1410/20141008tleBaldridge001h990.jpg',
-  'https://apod.nasa.gov/apod/image/1409/volcanicpillar_vetter_960.jpg',
-  'https://apod.nasa.gov/apod/image/1409/m27_snyder_960.jpg',
-  'https://apod.nasa.gov/apod/image/1409/PupAmulti_rot0.jpg',
-  'https://apod.nasa.gov/apod/image/1510/lunareclipse_27Sep_beletskycrop4.jpg',
-];
-
-type Props = $ReadOnly<{||}>;
-type State = {|likes: number|};
-class LikeCount extends React.Component<Props, State> {
-  state = {
-    likes: 7,
-  };
-
-  onClick = () => {
-    this.setState({likes: this.state.likes + 1});
-  };
-
-  render() {
-    const thumbsUp = '\uD83D\uDC4D';
-    return (
-      <View style={styles.likeContainer}>
-        <TouchableOpacity onPress={this.onClick} style={styles.likeButton}>
-          <Text style={styles.likesText}>{thumbsUp + ' Like'}</Text>
-        </TouchableOpacity>
-        <Text style={styles.likesText}>{this.state.likes + ' likes'}</Text>
-      </View>
-    );
-  }
-}
-
-class Button extends React.Component {
-  _handlePress = () => {
-    if (this.props.enabled && this.props.onPress) {
-      this.props.onPress();
-    }
-  };
-
-  render() {
-    return (
-      <TouchableWithoutFeedback onPress={this._handlePress}>
-        <View
-          style={[
-            styles.button,
-            this.props.enabled ? {} : styles.buttonDisabled,
-          ]}>
-          <Text style={styles.buttonText}>{this.props.text}</Text>
-        </View>
-      </TouchableWithoutFeedback>
-    );
-  }
-}
-
-class ProgressBar extends React.Component {
-  render() {
-    const fractionalPosition =
-      this.props.progress.position + this.props.progress.offset;
-    const progressBarSize =
-      (fractionalPosition / (this.props.numberOfPages - 1)) * this.props.size;
-    return (
-      <View style={[styles.progressBarContainer, {width: this.props.size}]}>
-        <View style={[styles.progressBar, {width: progressBarSize}]} />
-      </View>
-    );
-  }
-}
+import { PAGES, BGCOLOR, IMAGE_URIS, createPage } from "./Common";
+import { Button } from "./src/component/Button";
+import { LikeCount } from "./src/component/LikeCount";
+import { ProgressBar } from "./src/component/ProgressBar";
 
 export default class ViewPagerAndroidExample extends React.Component {
   constructor(props) {
@@ -169,6 +102,13 @@ export default class ViewPagerAndroidExample extends React.Component {
 
   render() {
     const {page, pages, animationsAreEnabled} = this.state;
+    //This is not implemented on iOS
+    let addPageButton;
+    if (Platform.OS === 'ios') {
+      addPageButton = null
+    } else {
+      addPageButton =  <Button enabled={true} text="Add new page" onPress={this.addPage} />;
+    }
     return (
       <View style={styles.container}>
         <ViewPagerAndroid
@@ -194,11 +134,7 @@ export default class ViewPagerAndroidExample extends React.Component {
               this.setState({scrollEnabled: !this.state.scrollEnabled})
             }
           />
-          <Button
-            enabled={true}
-            text="Add new page"
-            onPress={this.addPage}
-          />
+          {addPageButton}
         </View>
         <View style={styles.buttons}>
           {animationsAreEnabled ? (
@@ -253,18 +189,6 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'space-between',
   },
-  button: {
-    flex: 1,
-    width: 0,
-    margin: 5,
-    borderColor: 'gray',
-    borderWidth: 1,
-    backgroundColor: 'gray',
-  },
-  buttonDisabled: {
-    backgroundColor: 'black',
-    opacity: 0.5,
-  },
   buttonText: {
     color: 'white',
   },
@@ -279,34 +203,6 @@ const styles = StyleSheet.create({
     width: 300,
     height: 200,
     padding: 20,
-  },
-  likeButton: {
-    backgroundColor: 'rgba(0, 0, 0, 0.1)',
-    borderColor: '#333333',
-    borderWidth: 1,
-    borderRadius: 5,
-    flex: 1,
-    margin: 8,
-  },
-  likeContainer: {
-    flexDirection: 'row',
-    height: 45,
-  },
-  likesText: {
-    flex: 1,
-    fontSize: 18,
-    alignSelf: 'center',
-  },
-  progressBarContainer: {
-    height: 10,
-    margin: 10,
-    borderColor: '#eeeeee',
-    borderWidth: 2,
-  },
-  progressBar: {
-    alignSelf: 'flex-start',
-    flex: 1,
-    backgroundColor: '#eeeeee',
   },
   viewPager: {
     flex: 1,
