+++ example/index.js
@@ -4,12 +4,12 @@ import React from 'react';
 import {AppRegistry} from 'react-native';
 
 import {name as appName}  from './app.json';
-import ViewPagerAndroidExample from "./ViewPagerAndroidExample";
+import ViewPagerExample from "./ViewPagerExample";
 
 class ExampleApp extends React.Component<{}> {
   render() {
     return (
-        <ViewPagerAndroidExample />
+        <ViewPagerExample />
     )
   }
 }
